requirements.txt
flask==2.2.2
pandas==1.4.0
sqlalchemy==1.4.40
python-dotenv==0.21.0
psycopg2==2.9.3
psycopg2-binary== 2.9.3

